webapi/tct-extra-html5-tests/semantics/embedded-content-0/the-track-element/track_attributes.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xinx, liu <xinx.liu@intel.com>

-->

<html>
  <head>
    <title>track Test: track_attributes</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/embedded-content-0.html#the-track-element" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check that the track attributes are implemented as specified for existence, type and readonly" />
    <script src="../../../resources/testharness.js"></script>
    <script src="../../../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <video>
      <track id = "track">
    </video>
    <script type="text/javascript">
        var track = document.getElementById("track");
        [

            // HTMLTrackElement interface
            ["DOMString", "kind"],
            ["DOMString", "src"],
            ["DOMString", "srclang"],
            ["DOMString", "label"],
            ["boolean", "default"],
            ["unsigned short", "NONE", 0],
            ["unsigned short", "LOADING", 1],
            ["unsigned short", "LOADED", 2],
            ["unsigned short", "ERROR", 3],
            ["unsigned short", "readyState", "readonly"],
            ["object", "track", "readonly"],

        ].forEach(function(attr) {
            var type = attr[0];
            var name = attr[1];

            // existence check
            test(function() {
                assert_true(name in track, name + " attribute in track element");
            }, "Check if track." + name + " exists");

            // type, readonly and/or constant value check
            switch(type) {
                case "boolean":
                    test(function() {
                        assert_equals(typeof track[name], "boolean", name + " attribute of type");
                    }, "Check if track." + name + " is of type boolean");
                    break;

                case "DOMString":
                    test(function() {
                        assert_equals(typeof track[name], "string", name + " attribute of type");
                    }, "Check if track." + name + " is of type string");
                    break;

                case "unsigned short":
                    test(function() {
                        assert_equals(typeof track[name], "number", name + " attribute of type");
                    }, "Check if track." + name + " is of type number");
                    if(attr[2] != "test") {
                        test(function() {
                            assert_equals(track[name], attr[2], name + " attribute of value");
                        }, "Check if track." + name + " value is constant");
                    } else {
                        track[name] =attr[2];
                        test(function() {
                            assert_not_equals(track[name], attr[2], name + " attribute of value");
                        }, "Check if track." + name + " value is readonly");
                    }
                    break;

                case "object":
                    test(function() {
                        assert_equals(typeof track[name], "object", name + " attribute of type");
                    }, "Check if track." + name + " is of type object");
                    track[name] =attr[2];
                    test(function() {
                        assert_not_equals(track[name], attr[2], name + " attribute of value");
                    }, "Check if track." + name + " value is readonly");
                    break;

                default:
                    break;
            }
        });
    </script>
  </body>
</html>
